+++ python-backend/tests/mines/resources/test_mine_type_resource.py
@@ -45,7 +45,10 @@ def test_post_mine_disturbance_success(test_client, db_session, auth_headers):
     disturb = SampleMineDisturbanceCodes(mine_type.mine_tenure_type, 1)
     mine = MineFactory(mine_type=None)
 
-    test_data = {'mine_tenure_type_code': 'MIN', 'mine_disturbance_code': disturb}
+    test_data = {
+        'mine_tenure_type_code': mine_type.mine_tenure_type_code,
+        'mine_disturbance_code': disturb
+    }
     post_resp = test_client.post(f'/mines/{mine.mine_guid}/mine-types',
                                  json=test_data,
                                  headers=auth_headers['full_auth_header'])
@@ -70,7 +73,39 @@ def test_post_mine_type_missing_mine_tenure_type_code(test_client, db_session, a
 def test_post_mine_type_detail_invalid_mine_disturbance_code(test_client, db_session, auth_headers):
     mine_guid = MineFactory(mine_type=None).mine_guid
 
-    test_mine_type_data = {'mine_tenure_type_code': 'MIN', 'mine_disturbance_code': 'ABC'}
+    test_mine_type_data = {'mine_tenure_type_code': 'MIN', 'mine_disturbance_code': ['ABC']}
+    post_resp = test_client.post(f'/mines/{mine_guid}/mine-types',
+                                 json=test_mine_type_data,
+                                 headers=auth_headers['full_auth_header'])
+    post_data = json.loads(post_resp.data.decode())
+    assert post_resp.status_code == 400
+    assert post_data['message'] is not None
+
+
+def test_post_mine_type_detail_invalid_mine_commdity_code(test_client, db_session, auth_headers):
+    mine_guid = MineFactory(mine_type=None).mine_guid
+
+    test_mine_type_data = {'mine_tenure_type_code': 'MIN', 'mine_commodity_code': ['ZZZ']}
+    post_resp = test_client.post(f'/mines/{mine_guid}/mine-types',
+                                 json=test_mine_type_data,
+                                 headers=auth_headers['full_auth_header'])
+    post_data = json.loads(post_resp.data.decode())
+    assert post_resp.status_code == 400
+    assert post_data['message'] is not None
+
+
+def test_post_mine_type_detail_invalid_mine_commdity_code(test_client, db_session, auth_headers):
+    mine_guid = MineFactory(mine_type=None).mine_guid
+    mine_type = MineTypeFactory(mine_tenure_type_code='MIN',
+                                mine_type_detail={
+                                    'commodities': 0,
+                                    'disturbances': 0
+                                })
+    #make mine_type with cole but commodity only valid with Mineral
+    test_mine_type_data = {
+        'mine_tenure_type_code': 'COL',
+        'mine_commodity_code': SampleMineCommodityCodes(mine_type.mine_tenure_type, 1)[0]
+    }
     post_resp = test_client.post(f'/mines/{mine_guid}/mine-types',
                                  json=test_mine_type_data,
                                  headers=auth_headers['full_auth_header'])
@@ -108,6 +143,23 @@ def test_post_mine_type_duplicate(test_client, db_session, auth_headers):
                                   headers=auth_headers['full_auth_header'])
     post_data2 = json.loads(post_resp2.data.decode())
     assert post_resp2.status_code == 400
+
+
+def test_post_mine_type_detail_commodity_and_disturbance(test_client, db_session, auth_headers):
+    mine_guid = MineFactory(mine_type=None).mine_guid
+    mine_type = MineTypeFactory(mine_type_detail={'commodities': 0, 'disturbances': 0})
+
+    test_data = {
+        'mine_tenure_type_code': 'MIN',
+        'mine_disturbance_code': SampleMineDisturbanceCodes(mine_type.mine_tenure_type, 1)[0],
+        'mine_commodity_code': SampleMineCommodityCodes(mine_type.mine_tenure_type, 1)[0]
+    }
+    post_resp = test_client.post(f'/mines/{mine_guid}/mine-types',
+                                 json=test_data,
+                                 headers=auth_headers['full_auth_header'])
+    post_data = json.loads(post_resp.data.decode())
+    assert post_resp.status_code == 400
+    assert post_data['message'] is not None
 
 
 # DELETE
